refactor(utils): Split up navigator into several smaller classes
Move common functions into QkNavigator, and move external navigations to ExternalNavigator
