+++ presentation/src/main/java/com/moez/QKSMS/feature/compose/ComposeViewModel.kt
@@ -39,6 +39,7 @@ import com.moez.QKSMS.manager.MediaRecorderManager.AUDIO_FILE_SUFFIX
 import com.uber.autodispose.android.lifecycle.scope
 import com.uber.autodispose.autoDisposable
 import dev.octoshrimpy.quik.R
+import dev.octoshrimpy.quik.common.ExternalNavigator
 import dev.octoshrimpy.quik.common.Navigator
 import dev.octoshrimpy.quik.common.base.QkViewModel
 import dev.octoshrimpy.quik.common.util.ClipboardUtils
@@ -120,6 +121,7 @@ class ComposeViewModel @Inject constructor(
     private val messageRepo: MessageRepository,
     private val scheduledMessageRepo: ScheduledMessageRepository,
     private val navigator: Navigator,
+    private val externalNavigator: ExternalNavigator,
     private val permissionManager: PermissionManager,
     private val phoneNumberUtils: PhoneNumberUtils,
     private val prefs: Preferences,
@@ -401,7 +403,7 @@ class ComposeViewModel @Inject constructor(
                     ?: conversation.recipients.firstOrNull()?.address  // first recipient in convo
             }
             .autoDisposable(view.scope())
-            .subscribe { navigator.makePhoneCall(it) }
+            .subscribe { externalNavigator.makePhoneCall(it) }
 
         // Open the conversation settings if info button is clicked
         view.optionsItemIntent
@@ -1194,7 +1196,7 @@ class ComposeViewModel @Inject constructor(
                 }
 
                 if ((delay != 0 || state.scheduled != 0L) && !permissionManager.hasExactAlarms()) {
-                    navigator.showExactAlarmsSettings()
+                    externalNavigator.showExactAlarmsSettings()
                     return@withLatestFrom false
                 }
