+++ presentation/src/main/java/com/moez/QKSMS/feature/main/MainViewModel.kt
@@ -22,6 +22,7 @@ import androidx.recyclerview.widget.ItemTouchHelper
 import com.uber.autodispose.android.lifecycle.scope
 import com.uber.autodispose.autoDisposable
 import dev.octoshrimpy.quik.R
+import dev.octoshrimpy.quik.common.ExternalNavigator
 import dev.octoshrimpy.quik.common.Navigator
 import dev.octoshrimpy.quik.common.base.QkViewModel
 import dev.octoshrimpy.quik.extensions.mapNotNull
@@ -78,6 +79,7 @@ class MainViewModel @Inject constructor(
     private val markUnread: MarkUnread,
     private val speakThreads: SpeakThreads,
     private val navigator: Navigator,
+    private val externalNavigator: ExternalNavigator,
     private val permissionManager: PermissionManager,
     private val prefs: Preferences,
     private val ratingManager: RatingManager,
@@ -250,7 +252,7 @@ class MainViewModel @Inject constructor(
 
         view.changelogMoreIntent
                 .autoDisposable(view.scope())
-                .subscribe { navigator.showChangelog() }
+                .subscribe { externalNavigator.showChangelog() }
 
         view.queryChangedIntent
                 .debounce(200, TimeUnit.MILLISECONDS)
@@ -336,7 +338,7 @@ class MainViewModel @Inject constructor(
                         NavItem.ABOUT -> navigator.showAbout()
 //                        NavItem.PLUS -> navigator.showQksmsPlusActivity("main_menu")
 //                        NavItem.HELP -> navigator.showSupport()
-                        NavItem.INVITE -> navigator.showInvite()
+                        NavItem.INVITE -> externalNavigator.showInvite()
                         else -> Unit
                     }
                     drawerItem
@@ -396,7 +398,7 @@ class MainViewModel @Inject constructor(
                 .mapNotNull(conversationRepo::getConversation)
                 .map { conversation -> conversation.recipients }
                 .mapNotNull { recipients -> recipients[0]?.address?.takeIf { recipients.size == 1 } }
-                .doOnNext(navigator::addContact)
+                .doOnNext(externalNavigator::addContact)
                 .autoDisposable(view.scope())
                 .subscribe()
 
@@ -464,7 +466,7 @@ class MainViewModel @Inject constructor(
         view.rateIntent
                 .autoDisposable(view.scope())
                 .subscribe {
-                    navigator.showRating()
+                    externalNavigator.showRating()
                     ratingManager.rate()
                 }
 
@@ -553,7 +555,7 @@ class MainViewModel @Inject constructor(
                                     ?.address // most recent non-me msg address
                                 ?: conversationRepo.getConversation(threadId)
                                     ?.recipients?.firstOrNull()?.address  // first recipient in convo
-                            )?.let(navigator::makePhoneCall)
+                            )?.let(externalNavigator::makePhoneCall)
                         }
                         Preferences.SWIPE_ACTION_READ -> markRead.execute(listOf(threadId))
                         Preferences.SWIPE_ACTION_UNREAD -> markUnread.execute(listOf(threadId))
@@ -576,7 +578,7 @@ class MainViewModel @Inject constructor(
                         !state.contactPermission -> view.requestPermissions()
                         !state.notificationPermission -> {
                             if (prefs.hasAskedForNotificationPermission.get()) {
-                                navigator.showPermissions()
+                                externalNavigator.showPermissionsInSettings()
                             } else {
                                 prefs.hasAskedForNotificationPermission.set(true)
                                 view.requestPermissions()
